misc. adjustments
