+++ sale_order_payment_charge_line/README.rst
@@ -1,6 +1,6 @@
-===========================
-Website Sale Payment Charge
-===========================
+==============================
+Sale Order Payment Charge Line
+==============================
 
 .. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
    !! This file is generated by oca-gen-addon-readme !!
@@ -14,16 +14,17 @@ Website Sale Payment Charge
     :target: http://www.gnu.org/licenses/lgpl-3.0-standalone.html
     :alt: License: LGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-qrtl%2Fmi7--custom-lightgray.png?logo=github
-    :target: https://github.com/qrtl/mi7-custom/tree/10.0/website_sale_payment_charge
+    :target: https://github.com/qrtl/mi7-custom/tree/10.0/sale_order_payment_charge_line
     :alt: qrtl/mi7-custom
 
 |badge1| |badge2| |badge3| 
 
 This module does the following:
 
-- Adds a is_payment_charge field in product template.
-- Adds a is_payment_charge field in sale order line tree and form.
-  This field related to product_template.is_payment_charge.
+- Adds is_payment_charge field in product.template and sale.order.line.
+
+The module does not contain any business logic per se, but is expected to be depended
+upon by other modules with business logic.
 
 **Table of contents**
 
@@ -36,7 +37,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/qrtl/mi7-custom/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/qrtl/mi7-custom/issues/new?body=module:%20website_sale_payment_charge%0Aversion:%2010.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/qrtl/mi7-custom/issues/new?body=module:%20sale_order_payment_charge_line%0Aversion:%2010.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -51,6 +52,6 @@ Authors
 Maintainers
 ~~~~~~~~~~~
 
-This module is part of the `qrtl/mi7-custom <https://github.com/qrtl/mi7-custom/tree/10.0/website_sale_payment_charge>`_ project on GitHub.
+This module is part of the `qrtl/mi7-custom <https://github.com/qrtl/mi7-custom/tree/10.0/sale_order_payment_charge_line>`_ project on GitHub.
 
 You are welcome to contribute.
