+++ sale_order_payment_charge_line/__manifest__.py
@@ -2,13 +2,13 @@
 # Copyright 2021 Quartile Limited
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
 {
-    "name": "Website Sale Payment Charge",
+    "name": "Sale Order Payment Charge Line",
     "version": "10.0.1.0.0",
     "category": "Sale",
     "license": "LGPL-3",
     "author": "Quartile Limited",
     "website": "https://www.quartile.co",
-    "depends": ["website_sale"],
-    "data": ["views/product_template_view.xml", "views/sale_order_line_view.xml"],
+    "depends": ["sale"],
+    "data": ["views/product_template_views.xml"],
     "installable": True,
 }
