+++ sale_order_payment_charge_line/i18n/ja.po
@@ -1,0 +1,41 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* sale_order_payment_charge_line
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 10.0+e\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2021-11-11 15:20+0000\n"
+"PO-Revision-Date: 2021-11-11 15:20+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: sale_order_payment_charge_line
+#: model:ir.model.fields,field_description:sale_order_payment_charge_line.field_product_product_is_payment_charge
+#: model:ir.model.fields,field_description:sale_order_payment_charge_line.field_product_template_is_payment_charge
+#: model:ir.model.fields,field_description:sale_order_payment_charge_line.field_sale_order_line_is_payment_charge
+msgid "Payment Charge"
+msgstr "支払手数料"
+
+#. module: sale_order_payment_charge_line
+#: model:ir.model,name:sale_order_payment_charge_line.model_product_template
+msgid "Product Template"
+msgstr "プロダクトテンプレート"
+
+#. module: sale_order_payment_charge_line
+#: model:ir.model,name:sale_order_payment_charge_line.model_sale_order_line
+msgid "Sales Order Line"
+msgstr "販売オーダ明細"
+
+#. module: sale_order_payment_charge_line
+#: model:ir.model.fields,help:sale_order_payment_charge_line.field_product_product_is_payment_charge
+#: model:ir.model.fields,help:sale_order_payment_charge_line.field_product_template_is_payment_charge
+#: model:ir.model.fields,help:sale_order_payment_charge_line.field_sale_order_line_is_payment_charge
+msgid "Select this for payment charge products such as cash on delivery fees."
+msgstr "代引手数料等の支払手数料プロダクトにつきこの項目を選択してください。"
+
