+++ sale_order_payment_charge_line/models/product_template.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+# Copyright 2021 Quartile Limited
+# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
+
+from odoo import fields, models
+
+
+class ProductTemplate(models.Model):
+    _inherit = "product.template"
+
+    is_payment_charge = fields.Boolean(
+        "Payment Charge",
+        help="Select this for payment charge products such as cash on delivery fees.",
+    )
