+++ sale_order_payment_charge_line/readme/DESCRIPTION.rst
@@ -1,0 +1,6 @@
+This module does the following:
+
+- Adds is_payment_charge field in product.template and sale.order.line.
+
+The module does not contain any business logic per se, but is expected to be depended
+upon by other modules with business logic.
