+++ sale_order_payment_charge_line/static/description/index.html
@@ -4,7 +4,7 @@
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
 <meta name="generator" content="Docutils 0.15.1: http://docutils.sourceforge.net/" />
-<title>Website Sale Payment Charge</title>
+<title>Sale Order Payment Charge Line</title>
 <style type="text/css">
 
 /*
@@ -360,20 +360,20 @@ ul.auto-toc {
 </style>
 </head>
 <body>
-<div class="document" id="website-sale-payment-charge">
-<h1 class="title">Website Sale Payment Charge</h1>
+<div class="document" id="sale-order-payment-charge-line">
+<h1 class="title">Sale Order Payment Charge Line</h1>
 
 <!-- !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !! This file is generated by oca-gen-addon-readme !!
 !! changes will be overwritten.                   !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
-<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/lgpl-3.0-standalone.html"><img alt="License: LGPL-3" src="https://img.shields.io/badge/licence-LGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/qrtl/mi7-custom/tree/10.0/website_sale_payment_charge"><img alt="qrtl/mi7-custom" src="https://img.shields.io/badge/github-qrtl%2Fmi7--custom-lightgray.png?logo=github" /></a></p>
+<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/lgpl-3.0-standalone.html"><img alt="License: LGPL-3" src="https://img.shields.io/badge/licence-LGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/qrtl/mi7-custom/tree/10.0/sale_order_payment_charge_line"><img alt="qrtl/mi7-custom" src="https://img.shields.io/badge/github-qrtl%2Fmi7--custom-lightgray.png?logo=github" /></a></p>
 <p>This module does the following:</p>
 <ul class="simple">
-<li>Adds a is_payment_charge field in product template.</li>
-<li>Adds a is_payment_charge field in sale order line tree and form.
-This field related to product_template.is_payment_charge.</li>
+<li>Adds is_payment_charge field in product.template and sale.order.line.</li>
 </ul>
+<p>The module does not contain any business logic per se, but is expected to be depended
+upon by other modules with business logic.</p>
 <p><strong>Table of contents</strong></p>
 <div class="contents local topic" id="contents">
 <ul class="simple">
@@ -390,7 +390,7 @@ This field related to product_template.is_payment_charge.</li>
 <p>Bugs are tracked on <a class="reference external" href="https://github.com/qrtl/mi7-custom/issues">GitHub Issues</a>.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-<a class="reference external" href="https://github.com/qrtl/mi7-custom/issues/new?body=module:%20website_sale_payment_charge%0Aversion:%2010.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
+<a class="reference external" href="https://github.com/qrtl/mi7-custom/issues/new?body=module:%20sale_order_payment_charge_line%0Aversion:%2010.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
 <p>Do not contact contributors directly about support or help with technical issues.</p>
 </div>
 <div class="section" id="credits">
@@ -403,7 +403,7 @@ If you spotted it first, help us smashing it by providing a detailed and welcome
 </div>
 <div class="section" id="maintainers">
 <h2><a class="toc-backref" href="#id4">Maintainers</a></h2>
-<p>This module is part of the <a class="reference external" href="https://github.com/qrtl/mi7-custom/tree/10.0/website_sale_payment_charge">qrtl/mi7-custom</a> project on GitHub.</p>
+<p>This module is part of the <a class="reference external" href="https://github.com/qrtl/mi7-custom/tree/10.0/sale_order_payment_charge_line">qrtl/mi7-custom</a> project on GitHub.</p>
 <p>You are welcome to contribute.</p>
 </div>
 </div>
