+++ sale_order_payment_charge_line/views/product_template_views.xml
@@ -3,11 +3,13 @@
     <record id="product_template_form_view" model="ir.ui.view">
         <field name="name">product.template.common.form</field>
         <field name="model">product.template</field>
-        <field name="inherit_id" ref="sale.product_template_form_view_invoice_policy" />
+        <field name="inherit_id" ref="product.product_template_form_view" />
         <field name="arch" type="xml">
-            <field name="invoice_policy" position="after">
-                <field name="is_payment_charge" />
-            </field>
+            <xpath expr="//page[@name='general_information']/group[1]" position="inside">
+                <group>
+                    <field name="is_payment_charge" />
+                </group>
+            </xpath>
         </field>
     </record>
 </odoo>
