--- website_sale_payment_charge/i18n/ja_JP.po
@@ -1,47 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#	* website_sale_payment_charge
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 10.0+e\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2021-11-11 05:47+0000\n"
-"PO-Revision-Date: 2021-11-11 05:47+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: website_sale_payment_charge
-#: model:ir.model.fields,help:website_sale_payment_charge.field_sale_order_line_is_payment_charge
-msgid "If this field is checked, this product is eligible for cash on delivery."
-msgstr "If this field is checked, this product is eligible for cash on delivery."
-
-#. module: website_sale_payment_charge
-#: model:ir.model.fields,help:website_sale_payment_charge.field_delivery_carrier_is_payment_charge
-#: model:ir.model.fields,help:website_sale_payment_charge.field_product_product_is_payment_charge
-#: model:ir.model.fields,help:website_sale_payment_charge.field_product_template_is_payment_charge
-msgid "If this product is eligible for cash on delivery, please check here."
-msgstr "If this product is eligible for cash on delivery, please check here."
-
-#. module: website_sale_payment_charge
-#: model:ir.model.fields,field_description:website_sale_payment_charge.field_delivery_carrier_is_payment_charge
-#: model:ir.model.fields,field_description:website_sale_payment_charge.field_product_product_is_payment_charge
-#: model:ir.model.fields,field_description:website_sale_payment_charge.field_product_template_is_payment_charge
-#: model:ir.model.fields,field_description:website_sale_payment_charge.field_sale_order_line_is_payment_charge
-msgid "Is Payment Charge."
-msgstr "決済手数料"
-
-#. module: website_sale_payment_charge
-#: model:ir.model,name:website_sale_payment_charge.model_product_template
-msgid "Product Template"
-msgstr "製品テンプレート"
-
-#. module: website_sale_payment_charge
-#: model:ir.model,name:website_sale_payment_charge.model_sale_order_line
-msgid "Sales Order Line"
-msgstr "受注明細"
-
