--- website_sale_payment_charge/readme/DESCRIPTION.rst
@@ -1,5 +1,0 @@
-This module does the following:
-
-- Adds a is_payment_charge field in product template.
-- Adds a is_payment_charge field in sale order line tree and form.
-  This field related to product_template.is_payment_charge.
